SEE COMMIT NOTES
Complete code cleanup, moved/ renamed files to match advanced template code, added a project file, added sln project file
