--- Properties/AssemblyInfo.cs
@@ -1,19 +1,0 @@
-using System.Reflection;
-using System.Runtime.InteropServices;
-using MelonLoader;
-//using BuildInfo = SmallFixToScenes.BuildInfo;
-
-[assembly: ComVisible(false)]
-
-[assembly: AssemblyTitle(BuildInfo.Name)]
-[assembly: AssemblyDescription(BuildInfo.Description)]
-[assembly: AssemblyCompany(BuildInfo.Company)]
-[assembly: AssemblyProduct(BuildInfo.Name)]
-[assembly: AssemblyCopyright("Copyright © " + BuildInfo.Author)]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-[assembly: AssemblyVersion(BuildInfo.Version)]
-[assembly: AssemblyFileVersion(BuildInfo.Version)]
-[assembly: MelonInfo(typeof(SmallFixToScenes.Implementation), BuildInfo.Name, BuildInfo.Version, BuildInfo.Author)]
-[assembly: MelonGame("Hinterland", "TheLongDark")]
