+++ SmallFixToScenes.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33801.468
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SmallFixToScenes", "VisualStudio\SmallFixToScenes.csproj", "{A6F88FA4-B767-44D6-A1D3-A5AB9F0C6FF4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A6F88FA4-B767-44D6-A1D3-A5AB9F0C6FF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A6F88FA4-B767-44D6-A1D3-A5AB9F0C6FF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A6F88FA4-B767-44D6-A1D3-A5AB9F0C6FF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A6F88FA4-B767-44D6-A1D3-A5AB9F0C6FF4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B85E9811-2F9F-45BE-8844-7812404DA2EB}
+	EndGlobalSection
+EndGlobal
