+++ VisualStudio/BuildInfo.cs
@@ -1,0 +1,33 @@
+﻿namespace SmallFixToScenes
+{
+    public static class BuildInfo
+    {
+        #region Mandatory
+        /// <summary>The machine readable name of the mod (no special characters or spaces)</summary>
+        public const string Name            = "SmallFixToScenes";
+        /// <summary>Who made the mod</summary>
+        public const string Author          = "Phaedrus";
+        /// <summary>Current version (Using Major.Minor.Build) </summary>
+        public const string Version         = "1.0.0";
+        /// <summary>Name used on GUI's, like ModSettings</summary>
+        public const string GUIName         = "Small Fix to Scenes";
+        #endregion
+
+        #region Optional
+        /// <summary>What the mod does</summary>
+        public const string Description     = null;
+        /// <summary>Company that made it</summary>
+        public const string Company         = null;
+        /// <summary>A valid download link</summary>
+        public const string DownloadLink    = null;
+        /// <summary>Copyright info</summary>
+        public const string Copyright       = "Copyright ©  2023";
+        /// <summary>Trademark info</summary>
+        public const string Trademark       = null;
+        /// <summary>Product Name (Generally use the Name)</summary>
+        public const string Product         = "SmallFixToScenes";
+        /// <summary>Culture info</summary>
+        public const string Culture         = null;
+        #endregion
+    }
+}
