+++ VisualStudio/SmallFixToScenes.csproj
@@ -1,0 +1,66 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <!-- DO NOT CHANGE ANYTHING BETWEEN THESE (unless you know 100% what you are changing)-->
+  <!-- Further info: This defines critical project details, like the net version, C# language version, null enable and generate assembly info -->
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <LangVersion>Latest</LangVersion>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+  </PropertyGroup>
+  <!-- DO NOT CHANGE ANYTHING BETWEEN THESE -->
+
+  <!-- List of used references -->
+  <ItemGroup>
+		<!-- This is the base reference. MUST EXIST -->
+		<PackageReference Include="STBlade.Modding.TLD.Il2CppAssemblies.Windows"  Version="2.10.0"/>
+		<!-- This is only if you use the ModSettings -->
+		<PackageReference Include="STBlade.Modding.TLD.ModSettings"               Version="1.9.0"/>
+  </ItemGroup>
+
+  <!-- Most commonly used using's. Add to this so you dont need to define them in the cs files -->
+  <ItemGroup>
+    <Using Include="System.Reflection" />
+    <Using Include="System.Diagnostics.CodeAnalysis" />
+    <Using Include="Il2Cpp" />
+    <Using Include="HarmonyLib" />
+    <Using Include="MelonLoader" />
+    <Using Include="UnityEngine" />
+    <Using Include="ModSettings" />
+  </ItemGroup>
+
+  <PropertyGroup>
+    <TheLongDarkPath>O:\Steam\steamapps\common\TheLongDark</TheLongDarkPath> <!-- This should point to wherever your games primary EXE is -->
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <Title>SmallFixToScenes</Title> <!-- name of the mod -->
+    <Authors>Phaedrus</Authors> <!-- CURRENT author of the project -->
+    <Company>$(AssemblyName)</Company>
+    <Description>Fixes applied to some scenes</Description> <!-- what the mod does -->
+    <Copyright>MIT License</Copyright>
+    <PackageProjectUrl></PackageProjectUrl>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <RepositoryUrl>https://github.com/Phaedrus3/SmallFixToScenes</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>TheLongDark</PackageTags>
+    <NeutralLanguage>en</NeutralLanguage>
+    <AssemblyVersion>1.0.0</AssemblyVersion>
+    <FileVersion>1.0.0</FileVersion>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <DebugType>embedded</DebugType>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+    <DebugType>embedded</DebugType>
+  </PropertyGroup>
+
+  <!-- This is a build event. It will copy your dll to the currently defined TheLongDarkPath's Mods folder -->
+  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
+    <Exec Command="xcopy /y &quot;$(BaseOutputPath)\$(Configuration)\net6.0\$(AssemblyName).dll&quot; &quot;$(TheLongDarkPath)\Mods&quot;" />
+  </Target>
+
+</Project>
